Fix an issue where the section map is blank
If no new QSOs are entered, the section map will be blank because the qsos_by_section array is initialized, but the map is always redrawn (for grayline). Either qsos_by_section needs to be kept around or we have to always populate this so the unconditional section map redraw has valid data upon which to populate the map.