Add Headless service, sync name of RAM disk, updated docs a bit. (#37)
* Create n1mm_view_headless.service

Needed this since dashboard no longer has the code to generate the graphs. So I run dashboard and headless. it duplicates the work but does add to the modulatory since most people will not be using headless.

* Update INSTALL_RASPI.md

Add instructions to install the headless service

* Update INSTALL_RASPI.md

Fixed a typo in "your"

* Update INSTALL_RASPI.md

Added comment that bullseye also works.

* Update INSTALL_RASPI.md

* Update Configure_Apache.md

Updated to use /mnt/ramdisk to be consistent with the rest of the documentation and rpi_install.sh. I suspect this document was created before the rpi_install.