Final update ready for Pull Request to move config info to n1mm_view.ini Note the ini file can be in ONLY one of the following directories: ~/.config/n1mm_view.ini ~/n1mm_view.ini ~/n1mm_view/n1mm_view.ini
There is error handling if it is in more than one place
If one is a developer and wants to use git to pull the latest, I would recommend
NOT putting the config file in the script directory. I put it in the pi user's ~ dir
I also put the logo in the homedir (I added a new config that defaults to logo.png) and I set it
to my logo in the config directory

I also fixed an issue I saw in testing with N1MM where it sent a PSK QSO as a mode of PSK
rather than PSK31 so I added that as a new digital mode

Let me know how you like this one and let me know if anything needs to be reworked.