Convert to reading config from an INI file (#68)
* Added images to prevent new logos from showing as a changed file
Added ini file for new config support
Added variations of .db names

* Initial deploy of config module. Note thius does yet work as I need to move to my Pi. I cannot run this on my Mac.

* This allows the code to work.
I added a Singleton to the config object so init just runs once to load the files.
I removed the parameter of config.ini until I figure out how to add the parameter
 to the invocation of the class with the Singleton
I updated a few items that did not have config. prior to the field like EVENT_TIME and QTH_LONGITUDE
I updated code in headless.py to not try and save the images if no data is returned. This happens when there are no QSOS.
I updated many logging statementas to put the module in the name
  (I Have not looked to see if there is a way for force logging to do that on its own)

* Updated the rest of the config.py Config class to read all the values from the ini file
Changed the format of the log statement to add module::function
Removed the modules labels I added to dataaccess and headless and graphics before I realized I could add it to the log formatter.
Made the config object check the format of the date passed. It raises an exceptiona nd exits if this is not set properly.
   Note the Trackback makes it really hard to see the error
Last thing is to add the ini file name back as a parameter

* This file is ignored but I am trying to see if I can do an add, then not have changes tracked.

* Added a comment to the logformatter in constants.py toindicate this is called
first and therefore sets the format for all subsequent logging

* Create config.ini

* Update config.ini

* Update .gitignore

* Delete config.ini

* Create config-sample.ini

* Update INSTALL_RASPI.md

Add comment about config-sample.ini

* Update rpi_install.sh

Copied the sample config file to its target directory.

* Updated config.py to add back Headless parameter as I omitted it.

* Update INSTALL_RASPI.md

* Update INSTALL_RASPI.md

Removed the reference to MyClubCall. I could not find an occurrence of that any more.

* Added updates to handle logging dynamically to set the log level from ini file
Added code to check for a valid ini file ane report where it was found
Added code to list th epossible places the ini file can be if not found or multiples found

* Updated init script output to go to the syslog file

* Final update ready for Pull Request to move config info to n1mm_view.ini
Note the ini file can be in ONLY one of the following directories:
~/.config/n1mm_view.ini
~/n1mm_view.ini
~/n1mm_view/n1mm_view.ini

There is error handling if it is in more than one place
If one is a developer and wants to use git to pull the latest, I would recommend
NOT putting the config file in the script directory. I put it in the pi user's ~ dir
I also put the logo in the homedir (I added a new config that defaults to logo.png) and I set it
to my logo in the config directory

I also fixed an issue I saw in testing with N1MM where it sent a PSK QSO as a mode of PSK
rather than PSK31 so I added that as a new digital mode

Let me know how you like this one and let me know if anything needs to be reworked.

* MAkes headless executable